replace invoke with an eventually that accepts config
